Create README.md
cs3520_final_project
====================

This is my final project "Surfers" in cs3520. The project was about online shoping webpage using CMS. The project was a prototype of imaginary surf shop. The webpage used video tag to embed a video file as background however, the video was overcapacity of github repo.  so In this git hub there is no video 

I tried to make a responsive web design without using bootstrap or zurb foundation.